chore: moved pages into routes/ folder and cleaned up import component line
