+++ app/_components/index.ts
@@ -1,0 +1,7 @@
+export { default as Button } from './button';
+export { default as Input } from './input';
+export { default as FloatingButton } from './floating-button';
+export { default as Item } from './item';
+export { default as Layout } from './layout';
+export { default as Message } from './message';
+export { default as TextArea } from './textarea';
